Further Vertex and Edge support
Added equality support for vertex and edge structs for convenience. Removed edge.incidence tuple and replaced with edge.vert1 and edge.vert2 as it was not convenient to use.  Added edge.IncidentOn to easily determine if an edge is incident to a vertex. Changed graph.adj to graph.incidence as it no longer holds adjacency information; rather it is now an incidence matrix.